+++ odoo/local-src/specific_stock/tests/__init__.py
@@ -6,3 +6,4 @@ from . import test_picking_transfer
 from . import test_product_category_expired_times
 from . import test_stock_inventory
 from . import test_time_limit_order
+from . import test_lot_voice_identifier